+++ Source/pfile.h
@@ -7,6 +7,7 @@
 
 #include "player.h"
 #include "DiabloUI/diabloui.h"
+#include "utils/mpq_writer.hpp"
 
 namespace devilution {
 
@@ -27,6 +28,7 @@ private:
 	bool clear_tables_;
 };
 
+MpqWriter &CurrentSaveArchive();
 const char *pfile_get_password();
 void pfile_write_hero(bool writeGameData = false, bool clearTables = !gbIsMultiplayer);
 bool pfile_ui_set_hero_infos(bool (*uiAddHeroInfo)(_uiheroinfo *));
